wip - moving to f# tests in a separate fsproj.  Hitting issues getting mocha to run the f# tests.  May witch to babel from webpack
