+++ .vscode/launch.json
@@ -44,13 +44,13 @@
             "request": "launch",
             "args": [
                 "--extensionDevelopmentPath=${workspaceFolder}${config:vsCodeExtensionPath}",
-                "--extensionTestsPath=${workspaceFolder}${config:vsCodeExtensionPath}/out/test/suite/index"
+                "--extensionTestsPath=${workspaceFolder}${config:vsCodeExtensionPath}/out/extension.js"
             ],
             "outFiles": [
                 "${workspaceFolder}${config:vsCodeExtensionPath}/out/**/*.js",
                 "${workspaceFolder}${config:vsCodeExtensionPath}/dist/**/*.js"
-            ],
-            "preLaunchTask": "tasks: watch-tests"
+            ]
+            //"preLaunchTask": "tasks: watch-build vsCode extension tests"
         }
     ]
 }
