+++ src/paket.lock
@@ -5,8 +5,36 @@ NUGET
     Expecto (9.0.4)
       FSharp.Core (>= 4.6)
       Mono.Cecil (>= 0.11.3)
+    Fable.Browser.Blob (1.1.4)
+      Fable.Core (>= 3.0)
+      FSharp.Core (>= 4.7.2)
+    Fable.Browser.Dom (2.6)
+      Fable.Browser.Blob (>= 1.1.4)
+      Fable.Browser.Event (>= 1.4.5)
+      Fable.Browser.WebStorage (>= 1.0.4)
+      Fable.Core (>= 3.0)
+      FSharp.Core (>= 4.7.2)
+    Fable.Browser.Event (1.4.6)
+      Fable.Browser.Gamepad (>= 1.0.3)
+      Fable.Core (>= 3.0)
+      FSharp.Core (>= 4.7.2)
+    Fable.Browser.Gamepad (1.0.3)
+      Fable.Core (>= 3.0)
+      FSharp.Core (>= 4.7.2)
+    Fable.Browser.WebStorage (1.0.4)
+      Fable.Browser.Event (>= 1.4.4)
+      Fable.Core (>= 3.0)
+      FSharp.Core (>= 4.7.2)
     Fable.Core (3.4)
       FSharp.Core (>= 4.7.2)
+    Fable.Electron (6.0)
+      Fable.Browser.Dom (>= 1.0)
+      Fable.Core (>= 3.0)
+      Fable.Node (>= 1.0)
+      FSharp.Core (>= 4.5.2)
+    Fable.Mocha (2.11)
+      Fable.Core (>= 3.0)
+      FSharp.Core (>= 4.7)
     Fable.Node (1.2)
       Fable.Core (>= 3.1.2)
       FSharp.Core (>= 5.0)
