+++ src/vscode/ubictionary/package.json
@@ -10,7 +10,7 @@
 		"Other"
 	],
 	"activationEvents": [
-        "*"
+		"*"
 	],
 	"main": "./dist/extension.js",
 	"contributes": {
@@ -26,21 +26,24 @@
 		"compile": "fable src && webpack",
 		"watch": "fable watch src --runWatch webpack",
 		"package": "fable src && webpack --mode production --devtool hidden-source-map",
-		"compile-tests": "# tsc -p . --outDir out # TODO: switch to fable",
-		"watch-tests": "# tsc -p . -w --outDir out # TODO: switch to fable",
-		"pretest": "npm run compile-tests && npm run compile && npm run lint",
-		"test": "node ./out/test/runTest.js"
+		"compile-tests": "fable test && webpack -o out --entry ./test/index.js",
+		"watch-tests": "fable watch test --runWatch webpack -o out",
+		"pretest": "npm run compile-tests",
+		"test": "npm run pretest && node ./out/test/runTest.js"
 	},
 	"devDependencies": {
+		"@sinonjs/fake-timers": "^7.1.2",
 		"@types/vscode": "^1.61.0",
-		"@types/glob": "^7.1.4",
-		"@types/mocha": "^9.0.0",
-		"@types/node": "14.x",
-		"eslint": "^8.1.0",
+		"@vscode/test-electron": "^1.6.2",
 		"glob": "^7.1.7",
-		"mocha": "^9.1.3",
+		"jest": "^27.3.1",
+		"json-loader": "^0.5.7",
+		"mocha": "^8.3.2",
+		"prettier": "^2.4.1",
 		"webpack": "^5.52.1",
-		"webpack-cli": "^4.8.0",
-		"@vscode/test-electron": "^1.6.2"
+		"webpack-cli": "^4.8.0"
+	},
+	"dependencies": {
+		"react": "^17.0.2"
 	}
 }
