+++ src/vscode/ubictionary/test/paket.references
@@ -1,0 +1,1 @@
+Fable.Mocha
